Unit tests for admin

+++ src/main/java/com/softserve/itacademy/kek/services/impl/OrderServiceImpl.java
@@ -196,10 +196,12 @@ public class OrderServiceImpl implements IOrderService {
     public void deleteAll() throws OrderServiceException {
         logger.info("Delete all orders from DB");
         try {
-            orderRepository.deleteAll();
-            orderRepository.flush();
+            if (!(orderRepository.findAll().isEmpty())) {
+                orderRepository.deleteAll();
 
-            logger.debug("All orders was deleted from DB");
+                logger.debug("All orders was deleted from DB");
+            }
+            logger.debug("There is no orders in the DB");
         } catch (Exception ex) {
             logger.error("Error while deleting orders from DB: ", ex);
             throw new OrderServiceException("An error occurred while deleting orders", ex);
